Change weight back to Int and add input validation to weight and gradYear

+++ Adrenaline/Adrenaline/Views/SignupSequence/FirstOpenView.swift
@@ -32,7 +32,7 @@ struct Height: Hashable {
 }
 
 struct Weight: Hashable {
-    var weight: Double
+    var weight: Int
     var unit: WeightUnit
 }
